+++ CHANGELOG.md
@@ -11,6 +11,13 @@
     * Addition of is\_attending?, is\_not\_attending?, is\_invited? and is\_shown? methods
     * update(api, field\_id) method now updates only changed fields, becoming update(api)
   * Add :network Section type
+  * Add custom inspect methods for:
+    * Event::Attendance (replace event with event.id if present)
+    * Event::Column (replace event with event.id if present)
+    * FlexiRecord::Column (replace flexi\_record with flexi\_record.id)
+    * FlexiRecord::Data (replace flexi\_record with flexi\_record.id)
+    * Badge::Requirement (replace badge with badge.osm\_key)
+    * Badge::Data (replace badge with badge.osm\_key)
 
 ## Version 0.3.0